requirements.txt
annotated-types==0.6.0
anyio==4.2.0
bleach==6.1.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
docopt==0.6.2
fastapi==0.109.2
greenlet==3.0.3
h11==0.14.0
httptools==0.6.1
idna==3.6
kaggle==1.6.4
numpy==1.26.3
pandas==2.2.0
pipreqs==0.4.13
psycopg2-binary==2.9.9
pydantic==2.6.1
pydantic_core==2.16.2
python-dateutil==2.8.2
python-dotenv==1.0.1
python-slugify==8.0.3
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.25
starlette==0.36.3
text-unidecode==1.3
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.2.0
uvicorn==0.27.0.post1
uvloop==0.19.0
watchfiles==0.21.0
webencodings==0.5.1
websockets==12.0
yarg==0.1.9
